Fix Issue With Changing Grid Size With Slider
This explanation may not be necessary but it is notes for myself:

A simple issue but when adjusting thew slider (range input) the grid would
dissapear and no elements were being added to the DOM. However, dumb mistake
number one the slider value wasnt being passed to generateGrid() but the "sliderValue"
element was being passed. Second, in order for the grid to display correctly it must be
deleted and regenerated.